refactor3
