+++ pkg/state/stateprovide/types_test.go
@@ -132,3 +132,64 @@ func mockedStructs2(t *testing.T, addWsDescriptor bool) (*mockAppStructs, *mockV
 func mockedStructs(t *testing.T) (*mockAppStructs, *mockViewRecords) {
 	return mockedStructs2(t, true)
 }
+
+func TestBundle(t *testing.T) {
+	newKey := func(qname appdef.QName, id istructs.RecordID) (k istructs.IStateKeyBuilder) {
+		k = newMapKeyBuilder(sys.Storage_View, qname)
+		k.PutRecordID(sys.Storage_Record_Field_ID, id)
+		return
+	}
+	t.Run("put", func(t *testing.T) {
+		b := newBundle()
+
+		b.put(newKey(testRecordQName1, istructs.RecordID(1)), state.ApplyBatchItem{})
+		b.put(newKey(testRecordQName1, istructs.RecordID(1)), state.ApplyBatchItem{})
+		b.put(newKey(testRecordQName1, istructs.RecordID(2)), state.ApplyBatchItem{})
+		b.put(newKey(testRecordQName2, istructs.RecordID(1)), state.ApplyBatchItem{})
+
+		require.Equal(t, 3, b.size())
+	})
+	t.Run("get", func(t *testing.T) {
+		b := newBundle()
+		b.put(newKey(testRecordQName1, istructs.RecordID(1)), state.ApplyBatchItem{})
+
+		tests := []struct {
+			name string
+			key  istructs.IStateKeyBuilder
+			want bool
+		}{
+			{
+				name: "Should be ok",
+				key:  newKey(testRecordQName1, istructs.RecordID(1)),
+				want: true,
+			},
+			{
+				name: "Should be not ok",
+				key:  newKey(testRecordQName1, istructs.RecordID(2)),
+				want: false,
+			},
+		}
+		for _, test := range tests {
+			_, ok := b.get(test.key)
+
+			require.Equal(t, test.want, ok)
+		}
+	})
+	t.Run("containsKeysForSameView", func(t *testing.T) {
+		require := require.New(t)
+		b := newBundle()
+
+		b.put(newKey(testRecordQName1, istructs.RecordID(1)), state.ApplyBatchItem{})
+		b.put(newKey(testRecordQName2, istructs.RecordID(2)), state.ApplyBatchItem{})
+		b.put(newKey(testRecordQName2, istructs.RecordID(3)), state.ApplyBatchItem{})
+		b.put(newKey(testRecordQName2, istructs.RecordID(4)), state.ApplyBatchItem{})
+		b.put(newKey(testRecordQName1, istructs.RecordID(5)), state.ApplyBatchItem{})
+
+		require.Equal(5, b.size(), "initial bundle size")
+		require.False(b.containsKeysForSameEntity(newMapKeyBuilder(sys.Storage_View, testRecordQName3)))
+
+		k := newMapKeyBuilder(sys.Storage_View, testRecordQName2)
+		require.True(b.containsKeysForSameEntity(k))
+		require.Equal(5, b.size(), "remain bundle size")
+	})
+}
