+++ pkg/sys/storages/impl_app_secrets_storage_test.go
@@ -5,12 +5,12 @@
 package storages
 
 import (
-	"context"
 	"testing"
 
 	"github.com/stretchr/testify/require"
 	"github.com/voedger/voedger/pkg/appdef"
 	"github.com/voedger/voedger/pkg/isecrets"
+	"github.com/voedger/voedger/pkg/state"
 	"github.com/voedger/voedger/pkg/sys"
 )
 
@@ -20,24 +20,18 @@ func TestAppSecretsStorage_BasicUsage(t *testing.T) {
 	secretBody := `{"secret":"key"}`
 	sr := &isecrets.SecretReaderMock{}
 	sr.On("ReadSecret", secret).Return([]byte(secretBody), nil)
-	s := ProvideAsyncActualizerStateFactory()(context.Background(), nilAppStructsFunc, nil, nil, nil, sr, nil, nil, nil, 0, 0)
-	kb, err := s.KeyBuilder(sys.Storage_AppSecret, appdef.NullQName)
-	require.NoError(err)
+	storage := NewAppSecretsStorage(sr)
+	kb := storage.NewKeyBuilder(appdef.NullQName, nil)
 	kb.PutString(sys.Storage_AppSecretField_Secret, secret)
-
-	sv, err := s.MustExist(kb)
+	sv, err := storage.(state.IWithGet).Get(kb)
 	require.NoError(err)
-
 	require.Equal(secretBody, sv.AsString(""))
 }
 func TestAppSecretsStorage(t *testing.T) {
 	t.Run("Should return error when key invalid", func(t *testing.T) {
-		s := ProvideAsyncActualizerStateFactory()(context.Background(), nilAppStructsFunc, nil, nil, nil, nil, nil, nil, nil, 0, 0)
-		kb, err := s.KeyBuilder(sys.Storage_AppSecret, appdef.NullQName)
-		require.NoError(t, err)
-
-		_, err = s.MustExist(kb)
-
+		storage := NewAppSecretsStorage(nil)
+		kb := storage.NewKeyBuilder(appdef.NullQName, nil)
+		_, err := storage.(state.IWithGet).Get(kb)
 		require.ErrorContains(t, err, "secret name is not specified")
 	})
 }
