+++ pkg/sys/storages/impl_federation_blob_storage_test.go
@@ -5,12 +5,12 @@
 package storages
 
 import (
-	"context"
 	"testing"
 
 	"github.com/stretchr/testify/require"
 	"github.com/voedger/voedger/pkg/appdef"
 	"github.com/voedger/voedger/pkg/istructs"
+	"github.com/voedger/voedger/pkg/state"
 	"github.com/voedger/voedger/pkg/sys"
 )
 
@@ -26,11 +26,8 @@ func TestFederationBlobStorage_BasicUsage(t *testing.T) {
 		require.Equal(int64(1), blobId)
 		return buffer, nil
 	}
-	s := ProvideAsyncActualizerStateFactory()(context.Background(), mockedHostStateStructs, nil, SimpleWSIDFunc(istructs.WSID(1)), nil, nil, nil, nil, nil, 0, 0, WithFederationBlobHandler(federatioBlobHandler))
-
-	k, err := s.KeyBuilder(sys.Storage_FederationBlob, appdef.NullQName)
-	require.NoError(err)
-
+	storage := NewFederationBlobStorage(nil, nil, nil, nil, federatioBlobHandler)
+	k := storage.NewKeyBuilder(appdef.NullQName, nil)
 	k.PutString(sys.Storage_FederationBlob_Field_Owner, "owner")
 	k.PutString(sys.Storage_FederationBlob_Field_AppName, "appname")
 	k.PutInt64(sys.Storage_FederationBlob_Field_BlobID, 1)
@@ -38,7 +35,7 @@ func TestFederationBlobStorage_BasicUsage(t *testing.T) {
 	k.PutInt64(sys.Storage_FederationBlob_Field_WSID, 123)
 
 	readBytes := make([]byte, 0)
-	err = s.Read(k, func(_ istructs.IKey, value istructs.IStateValue) (err error) {
+	err := storage.(state.IWithRead).Read(k, func(_ istructs.IKey, value istructs.IStateValue) (err error) {
 		readBytes = append(readBytes, value.AsBytes(sys.Storage_FederationBlob_Field_Body)...)
 		return
 	})
