+++ pkg/sys/storages/impl_http_storage_test.go
@@ -5,7 +5,6 @@
 package storages
 
 import (
-	"context"
 	"fmt"
 	"io"
 	"net/http"
@@ -16,12 +15,13 @@ import (
 	"github.com/stretchr/testify/require"
 	"github.com/voedger/voedger/pkg/appdef"
 	"github.com/voedger/voedger/pkg/istructs"
+	"github.com/voedger/voedger/pkg/state"
 	"github.com/voedger/voedger/pkg/sys"
 )
 
 func TestHttpStorage_BasicUsage(t *testing.T) {
 	require := require.New(t)
-	s := ProvideAsyncActualizerStateFactory()(context.Background(), nilAppStructsFunc, nil, nil, nil, nil, nil, nil, nil, 0, 0)
+	storage := NewHttpStorage(nil)
 	ts := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
 		require.Equal(http.MethodPost, r.Method)
 		require.Equal("my-value", r.Header.Get("my-header"))
@@ -32,10 +32,7 @@ func TestHttpStorage_BasicUsage(t *testing.T) {
 		_, _ = w.Write([]byte(`{"hello":"storage"}`))
 	}))
 	defer ts.Close()
-
-	k, err := s.KeyBuilder(sys.Storage_Http, appdef.NullQName)
-	require.NoError(err)
-
+	k := storage.NewKeyBuilder(appdef.NullQName, nil)
 	k.PutString(sys.Storage_Http_Field_Url, ts.URL)
 	k.PutString(sys.Storage_Http_Field_Method, http.MethodPost)
 	k.PutString(sys.Storage_Http_Field_Header, "my-header: my-value")
@@ -43,7 +40,7 @@ func TestHttpStorage_BasicUsage(t *testing.T) {
 	k.PutBytes(sys.Storage_Http_Field_Body, []byte(`{"hello":"api"}`))
 	require.Equal(fmt.Sprintf(`%s %s {"hello":"api"}`, http.MethodPost, ts.URL), k.(fmt.Stringer).String())
 	var v istructs.IStateValue
-	err = s.Read(k, func(_ istructs.IKey, value istructs.IStateValue) (err error) {
+	err := storage.(state.IWithRead).Read(k, func(_ istructs.IKey, value istructs.IStateValue) (err error) {
 		v = value
 		return
 	})
@@ -57,28 +54,22 @@ func TestHttpStorage_BasicUsage(t *testing.T) {
 func TestHttpStorage_Timeout(t *testing.T) {
 	t.Run("Should panic when url not found", func(t *testing.T) {
 		require := require.New(t)
-		s := ProvideAsyncActualizerStateFactory()(context.Background(), nilAppStructsFunc, nil, nil, nil, nil, nil, nil, nil, 0, 0)
-		k, err := s.KeyBuilder(sys.Storage_Http, appdef.NullQName)
-		require.NoError(err)
-
-		err = s.Read(k, func(istructs.IKey, istructs.IStateValue) error { return nil })
+		storage := NewHttpStorage(nil)
+		k := storage.NewKeyBuilder(appdef.NullQName, nil)
+		err := storage.(state.IWithRead).Read(k, func(istructs.IKey, istructs.IStateValue) error { return nil })
 		require.ErrorIs(err, ErrNotFound)
 	})
 	t.Run("Should return error on timeout", func(t *testing.T) {
 		require := require.New(t)
-		s := ProvideAsyncActualizerStateFactory()(context.Background(), nilAppStructsFunc, nil, nil, nil, nil, nil, nil, nil, 0, 0)
+		storage := NewHttpStorage(nil)
 		ts := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
 			time.Sleep(time.Millisecond * 200)
 		}))
 		defer ts.Close()
-
-		k, err := s.KeyBuilder(sys.Storage_Http, appdef.NullQName)
-		require.NoError(err)
+		k := storage.NewKeyBuilder(appdef.NullQName, nil)
 		k.PutString(sys.Storage_Http_Field_Url, ts.URL)
 		k.PutInt64(sys.Storage_Http_Field_HTTPClientTimeoutMilliseconds, 100)
-
-		err = s.Read(k, func(istructs.IKey, istructs.IStateValue) error { return nil })
-
+		err := storage.(state.IWithRead).Read(k, func(istructs.IKey, istructs.IStateValue) error { return nil })
 		require.Error(err)
 	})
 }
@@ -90,9 +81,7 @@ func TestHttpStorage_NewKeyBuilder_should_refresh_key_builder(t *testing.T) {
 	k.PutString(sys.Storage_Http_Field_Method, http.MethodPost)
 	k.PutString(sys.Storage_Http_Field_Header, "my-header: my-value")
 	k.PutBytes(sys.Storage_Http_Field_Body, []byte(`{"hello":"api"}`))
-
 	hskb := s.NewKeyBuilder(appdef.NullQName, k).(*httpStorageKeyBuilder)
-
 	require.Equal(sys.Storage_Http, hskb.Storage())
 	require.Equal("", hskb.url)
 	require.Equal("", hskb.method)
@@ -104,8 +93,6 @@ func TestHttpStorageKeyBuilder_headers(t *testing.T) {
 	require := require.New(t)
 	k := &httpStorageKeyBuilder{headers: make(map[string]string)}
 	k.PutString(sys.Storage_Http_Field_Header, "key: hello:world")
-
 	headers := k.headers
-
 	require.Equal("hello:world", headers["key"])
 }
