+++ pkg/sys/storages/impl_querycontext_storage_test.go
@@ -5,38 +5,24 @@
 package storages
 
 import (
-	"context"
 	"testing"
 
 	"github.com/stretchr/testify/require"
 
 	"github.com/voedger/voedger/pkg/appdef"
 	"github.com/voedger/voedger/pkg/istructs"
+	"github.com/voedger/voedger/pkg/state"
 	"github.com/voedger/voedger/pkg/sys"
 )
 
 func TestQueryContextStorage(t *testing.T) {
 	wsid := istructs.NullWSID
 	arg := istructs.NewNullObject()
-
-	wsidFunc := SimpleWSIDFunc(wsid)
-
-	execQueryArgsFunc := func() istructs.PrepareArgs {
-		return istructs.PrepareArgs{
-			Workpiece:      nil,
-			ArgumentObject: arg,
-			WSID:           wsid,
-			Workspace:      nil,
-		}
-	}
+	wsidFunc := state.SimpleWSIDFunc(wsid)
 	argFunc := func() istructs.IObject { return arg }
-
-	s := ProvideQueryProcessorStateFactory()(context.Background(), nil, nil, wsidFunc, nil, nil, nil, nil, execQueryArgsFunc, argFunc, nil, nil, nil)
-
-	kb, err := s.KeyBuilder(sys.Storage_QueryContext, appdef.NullQName)
-	require.NoError(t, err)
-
-	v, err := s.MustExist(kb)
+	storage := NewQueryContextStorage(argFunc, wsidFunc)
+	kb := storage.NewKeyBuilder(appdef.NullQName, nil)
+	v, err := storage.(state.IWithGet).Get(kb)
 	require.NoError(t, err)
 	require.Equal(t, int64(wsid), v.AsInt64(sys.Storage_QueryContext_Field_Workspace))
 	require.NotNil(t, v.AsValue(sys.Storage_QueryContext_Field_ArgumentObject))
